% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspredicted_prereg.R
\name{aspredicted_prereg}
\alias{aspredicted_prereg}
\title{AsPredicted.org template}
\usage{
aspredicted_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document using the AsPredicted.org template
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "aspredicted_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}

