% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brandt_prereg.R
\name{brandt_prereg}
\alias{brandt_prereg}
\title{Brandt et al. Replication Recipe Preregistration template}
\usage{
brandt_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document using the questions suggested by Brandt et al. (2013)
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "brandt_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}
\references{
Brandt, M. J., IJzerman, H., Dijksterhuis, A., Farach, F. J., Geller, J., Giner-Sorolla, R., ... van 't Veer, A. (2014). The Replication Recipe: What makes for a convincing replication? Journal of Experimental Social Psychology, 50, 217--224. doi: \href{https://doi.org/10.1016/j.jesp.2013.10.005}{10.1016/j.jesp.2013.10.005}
}
