% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_prereg.R
\name{cos_prereg}
\alias{cos_prereg}
\title{COS Preregistration Challenge template
Knit a PDF document}
\usage{
cos_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{rmarkdown::pdf_document}}
}
\description{
COS Preregistration Challenge template
Knit a PDF document
}

