% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruewell_prereg.R
\name{cruewell_prereg}
\alias{cruewell_prereg}
\title{Crüwell & Evans preregistration for application of cognitive models}
\usage{
cruewell_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document using the Crüwell & Evans preregistration for application
of cognitive models
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "cruewell_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}
\references{
Crüwell, S. & Evans, N. J. (2021). Preregistration in diverse contexts: a preregistration template for the application of cognitive models. Royal Society Open Science. 8:210155 doi: \url{https://doi.org/10.1016/j.jesp.2013.10.005}
}
