% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_prereg.R
\name{fmri_prereg}
\alias{fmri_prereg}
\title{fmri preregistration template}
\usage{
fmri_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document using the fMRI preregistration template suggested by Jessica Flannery (https://osf.io/maqwc/).
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "fmri_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}
