% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg.R
\docType{package}
\name{prereg}
\alias{prereg}
\title{R Markdown Templates to Preregister Scientific Studies}
\description{
The provided \href{https://rmarkdown.rstudio.com/}{R Markdown} templates are based
on the \href{https://osf.io/x5w7h/}{Center for Open Science Preregistration Challenge},
the \href{https://aspredicted.org/}{AsPredicted.org} questions, a template suggested
for social psychology by van 't Veer and Giner-Sorolla (2016), and the replication recipe
suggested by Brandt et al. (2013).
}
\section{System requirements}{

Before using \pkg{prereg} to create a preregistration document, make sure the following
software is installed on your computer:

\itemize{
\item{\href{https://rstudio.com/}{RStudio} (>= 0.98.932); if you don't use RStudio
, you need to install \href{https://pandoc.org/}{pandoc} using the
\href{https://pandoc.org/installing.html}{instructions for your operating system}}
\item{A \href{https://de.wikipedia.org/wiki/TeX}{TeX} distribution (2013 or later; e.g., \href{https://miktex.org/}{MikTeX}
for Windows, \href{https://tug.org/mactex/}{MacTeX} for Mac, obviously, or \href{https://www.tug.org/texlive/}{TeX Live}
for Linux)}
}

If you are running \emph{Windows}, use MikTex if possible. Currently, pandoc and the Windows version of Tex Live
\href{https://github.com/rstudio/rmarkdown/issues/6}{don't seem to like each other}. Make sure you install the
\emph{complete}---not the basic---version.
}

\section{Author and Maintainer}{

Frederik Aust (frederik.aust at uni-koeln.de).
}

\references{
Brandt, M. J., IJzerman, H., Dijksterhuis, A., Farach, F. J., Geller, J., Giner-Sorolla, R., ... van 't Veer, A. (2014). The Replication Recipe: What makes for a convincing replication? Journal of Experimental Social Psychology, 50, 217--224. doi: \href{https://doi.org/10.1016/j.jesp.2013.10.005}{10.1016/j.jesp.2013.10.005}
van 't Veer, A. E., & Giner-Sorolla, R. (2016). Pre-registration in social psychology---A discussion and
suggested template. Journal of Experimental Social Psychology, 67, 2--12. doi:
\href{https://doi.org/10.1016/j.jesp.2016.03.004}{10.1016/j.jesp.2016.03.004}
}
