% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prp_quant_prereg.R
\name{prp_quant_prereg}
\alias{prp_quant_prereg}
\alias{psyquant_prereg}
\title{Psychological Research Preregistration-Quantitative (aka PRP-QUANT) Template}
\usage{
prp_quant_prereg(...)

psyquant_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document based on the Psychological Research Preregistration-Quantitative (aka PRP-QUANT) Template
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "prp_quant_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}
