% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psyquant_prereg.R
\name{psyquant_prereg}
\alias{psyquant_prereg}
\title{Preregistration for Quantitative Research in Psychology Template}
\usage{
psyquant_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document based on the Preregistration for Quantitative Research in Psychology Template
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "psyquant_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}
