% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_prereg.R
\name{rr_prereg}
\alias{rr_prereg}
\title{Registered Reports template}
\usage{
rr_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document for preregistration of Stage 1 Registered Reports
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "rr_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}
