% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vantveer_prereg.R
\name{vantveer_prereg}
\alias{vantveer_prereg}
\title{van 't Veer & Giner-Sorolla Preregistration template}
\usage{
vantveer_prereg(...)
}
\arguments{
\item{...}{additional arguments to \code{\link[rmarkdown]{pdf_document}}; \code{template} is ignored.}
}
\description{
Knit a PDF document using the questions suggested by van 't Veer and Giner-Sorolla (2016) for social
psychologists
}
\examples{
\dontrun{
# Create R Markdown file
rmarkdown::draft(
  "my_preregistration.Rmd"
  , "vantveer_prereg"
  , package = "prereg"
  , create_dir = FALSE
  , edit = FALSE
  )

# Render file
rmarkdown::render("my_preregistration.Rmd")
}

}
\references{
van 't Veer, A. E., & Giner-Sorolla, R. (2016). Pre-registration in social psychology---A discussion and
suggested template. Journal of Experimental Social Psychology, 67, 2--12. doi:
\href{https://doi.org/10.1016/j.jesp.2016.03.004}{10.1016/j.jesp.2016.03.004}
}
