% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg.R
\docType{package}
\name{prereg}
\alias{prereg}
\title{R Markdown Templates to Preregister Scientific Studies}
\description{
This package provides a collection of preregistration templates for empirical scientific studies.
}
\section{System requirements}{

\strong{prereg} depends on additional software, namely,
\enumerate{
\item \href{https://www.R-project.org/}{R} 2.11.1 or later and
\item \href{https://pandoc.org/}{pandoc} 1.19 or later
\item \href{https://en.wikipedia.org/wiki/TeX}{TeX} 2013 or later.
}

If you work with \href{https://www.rstudio.com/}{RStudio} (1.1.453 or later) pandoc should already be installed, otherwise refer to the \href{https://pandoc.org/installing.html}{installation instructions} for your operating system.

\strong{prereg} can be used with common TeX distributions, such as \href{http://miktex.org/}{MikTeX} on Windows, \href{https://tug.org/mactex/}{MacTeX} on Mac, or \href{http://www.tug.org/texlive/}{TeX Live} on Linux.

If you mainly use TeX to render R Markdown documents, we strongly recommend using the \href{https://yihui.org/tinytex/}{TinyTex} distribution.
It is lightweight and automatically installs missing LaTeX packages and can be installed from within R with \code{tinytex::install_tinytex()}.
}

\section{Author and Maintainer}{

Frederik Aust (frederik.aust at uni-koeln.de).
}

\references{
Bosnjak, M., Fiebach, C. J., Mellor, D., Mueller, S., O’Connor, D. B., Oswald, F. L., & Sokol-Chang, R. I. (2021). A template for preregistration of quantitative research in psychology: Report of the joint psychological societies preregistration task force. American Psychologist. doi: 10.1037/amp0000879

Brandt, M. J., IJzerman, H., Dijksterhuis, A., Farach, F. J., Geller, J., Giner-Sorolla, R., … van ’t Veer, A. (2014). The Replication Recipe: What makes for a convincing replication? Journal of Experimental Social Psychology, 50, 217–224. \doi{10.1016/j.jesp.2013.10.005}

Crüwell, S. & Evans, N. J. (2021). Preregistration in diverse contexts: a preregistration template for the application of cognitive models. Royal Society Open Science. 8:210155. \doi{10.1016/j.jesp.2013.10.005}

Flannery, J. E. (2020, October 22). fMRI Preregistration Template. Retrieved from \url{https://osf.io/6juft}

van ’t Veer, A. E., & Giner-Sorolla, R. (2016). Pre-registration in social psychology—A discussion and suggested template. Journal of Experimental Social Psychology, 67, 2–12. \doi{10.1016/j.jesp.2016.03.004}
}
